backend/asu_agent/agent.py
from google.adk.agents import Agent

root_agent = Agent(
    model="gemini-2.0-flash",
    name="asu_international_assistant",
    description="An assistant for international students at Arizona State University.",
    instruction=(
        "You are a helpful assistant for international students at Arizona State University. "
        "Give clear, practical guidance about visas (F-1, J-1, OPT, CPT), enrollment, "
        "housing, campus life, health insurance, and working on/off campus. "
        "When you are not certain, say so and point to official ASU or U.S. government sources."
    ),
)
