backend/requirements.txt
google-adk>=1.28
fastapi>=0.115
uvicorn[standard]>=0.30
pydantic>=2
